requirements.txt
astroid==2.4.2
astropy==4.0.1.post1
attrs==19.3.0
aws-cdk.assets==1.53.0
aws-cdk.aws-apigateway==1.53.0
aws-cdk.aws-applicationautoscaling==1.53.0
aws-cdk.aws-autoscaling-common==1.53.0
aws-cdk.aws-certificatemanager==1.53.0
aws-cdk.aws-cloudformation==1.53.0
aws-cdk.aws-cloudwatch==1.53.0
aws-cdk.aws-codeguruprofiler==1.53.0
aws-cdk.aws-dynamodb==1.53.0
aws-cdk.aws-ec2==1.53.0
aws-cdk.aws-efs==1.53.0
aws-cdk.aws-elasticloadbalancingv2==1.53.0
aws-cdk.aws-events==1.53.0
aws-cdk.aws-glue==1.53.0
aws-cdk.aws-iam==1.53.0
aws-cdk.aws-kinesis==1.53.0
aws-cdk.aws-kms==1.53.0
aws-cdk.aws-lakeformation==1.53.0
aws-cdk.aws-lambda==1.53.0
aws-cdk.aws-lambda-event-sources==1.53.0
aws-cdk.aws-logs==1.53.0
aws-cdk.aws-route53==1.53.0
aws-cdk.aws-s3==1.53.0
aws-cdk.aws-s3-assets==1.53.0
aws-cdk.aws-s3-notifications==1.53.0
aws-cdk.aws-sns==1.53.0
aws-cdk.aws-sns-subscriptions==1.53.0
aws-cdk.aws-sqs==1.53.0
aws-cdk.aws-ssm==1.53.0
aws-cdk.cloud-assembly-schema==1.53.0
aws-cdk.core==1.53.0
aws-cdk.custom-resources==1.53.0
aws-cdk.cx-api==1.53.0
aws-cdk.region-info==1.53.0
boto3==1.14.23
botocore==1.17.23
cattrs==1.0.0
constructs==3.0.4
docutils==0.15.2
isort==4.3.21
jmespath==0.10.0
jsii==1.9.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.0
publication==0.0.3
pylint==2.5.3
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
wrapt==1.12.1
